Add .bashrc

+++ dotcopy.py
@@ -1,17 +1,21 @@
+
 import sys
 import os
 import shutil
 import configparser
 
 
-def copy_files(source, destination):
+def copy_files(source, destination, is_file=False):
     source_path = os.path.expanduser(source)
     destination_path = os.path.expanduser(destination)
 
     try:
         if os.path.exists(destination_path):
             shutil.rmtree(destination_path)
-        shutil.copytree(source_path, destination_path)
+        if is_file:
+            shutil.copy2(source_path, destination_path)
+        else:
+            shutil.copytree(source_path, destination_path)
         print("Files copied successfully!")
     except Exception as e:
         print(f"An error occurred: {str(e)}")
@@ -36,13 +40,19 @@ def main():
     source_path = entry["source"]
     destination_path = entry["destination"]
 
+    if "file" in entry:
+        is_file = entry.getboolean("file")
+    else:
+        is_file = False
+
     if action == "put":
-        copy_files(source_path, destination_path)
+        copy_files(source_path, destination_path, is_file)
     elif action == "get":
-        copy_files(destination_path, source_path)
+        copy_files(destination_path, source_path, is_file)
     else:
         print("Invalid action. Use 'put' or 'get'.")
 
 
 if __name__ == "__main__":
     main()
+
